backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.1",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.11",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
